- add the AIXPatch (forgotten in the previous cvs commit) and a documentation about AIX installation - start looking at bosons on torus with magnetic translations - add new qhe hamiltonian of particles on sphere with laplacian delta interaction
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@25 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Hamiltonian/QHEHamiltonian/AbstractQHEHamiltonian.h
@@ -76,14 +76,14 @@ class AbstractQHEHamiltonian : public AbstractHamiltonian
   //
   // allowedMemory = amount of memory that cam be allocated for fast multiplication
   // return value = amount of memory needed
-  virtual int FastMultiplicationMemory(int allowedMemory) = 0;
+  virtual long FastMultiplicationMemory(long allowedMemory) = 0;
 
   // test the amount of memory needed for fast multiplication algorithm (partial evaluation)
   //
   // firstComponent = index of the first component that has to be precalcualted
   // lastComponent  = index of the last component that has to be precalcualted
   // return value = number of non-zero matrix element
-  virtual int PartialFastMultiplicationMemory(int firstComponent, int lastComponent) = 0;
+  virtual long PartialFastMultiplicationMemory(int firstComponent, int lastComponent) = 0;
 
   // enable fast multiplication algorithm
   //
