+++ src/Hamiltonian/QHEHamiltonian/ParticleOnSphereDeltaModifiedHamiltonian.h
@@ -248,7 +248,7 @@ class ParticleOnSphereDeltaModifiedHamiltonian : public AbstractHamiltonian
   // test the amount of memory needed for fast multiplication algorithm
   //
   // return value = amount of memory needed
-  int FastMultiplicationMemory();
+  long FastMultiplicationMemory();
 
   // enable fast multiplication algorithm
   //
