+++ src/Hamiltonian/QHEHamiltonian/ParticleOnTorusDeltaHamiltonian.h
@@ -62,7 +62,7 @@ class ParticleOnTorusDeltaHamiltonian : public AbstractQHEOnTorusHamiltonian
   // memory = maximum amount of memory that can be allocated for fast multiplication (negative if there is no limit)
   // precalculationFileName = option file name where precalculation can be read instead of reevaluting them
   ParticleOnTorusDeltaHamiltonian(ParticleOnTorus* particles, int nbrParticles, int maxMomentum, double ratio, 
-				  AbstractArchitecture* architecture, int memory = -1, char* precalculationFileName = 0);
+				  AbstractArchitecture* architecture, long memory = -1, char* precalculationFileName = 0);
 
   // destructor
   //
